requirements.txt
aioredis==1.3.1
asgi-redis==1.4.3
asgiref==3.2.0
async-timeout==3.0.1
attrs==19.3.0
autobahn==20.6.2
Automat==20.2.0
cffi==1.14.0
channels==2.4.0
channels-redis==3.0.0
constantly==15.1.0
cryptography==2.9.2
daphne==2.5.0
Django==3.0.8
django-channels-presence==1.0.0
hiredis==1.0.1
hyperlink==19.0.0
idna==2.10
incremental==17.5.0
msgpack==1.0.0
msgpack-python==0.5.6
pkg-resources==0.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyHamcrest==2.0.2
pyOpenSSL==19.1.0
pytz==2020.1
redis==2.10.6
service-identity==18.1.0
six==1.15.0
sqlparse==0.3.1
Twisted==20.3.0
txaio==20.4.1
zope.interface==5.1.0
